Clean up list styles.
This involves sweeping changes to code but very few actual display changes.
A few pages have changed in display for the sake of consistency.

Ironically, on a web site devoted to DeviCat, we are *removing* magic. And
it is a definite improvement. Less magic in web sites, more magic in art!