Switch completely to using emote detection.
This switches all the URLs to be v2, which shouldn't affect anything. It also
removes all the phantom emotes that have been annoying me for ages (yay!). It
removes a dependency, so now it's entirely built from that JSON file. And, if
that wasn't enough, that dep was the cause of five-second run times, so now
the entire run is done in a fraction of a second!